Move functions to reports data processor and remove the old draft/untested functions
